// server-nest/src/seeds/types/product-category.type.ts
// don't use this type in the main codebase
// it's only meant to be used for DB seeding
export type ProductCategory =
    | 'Ігрові консолі'
    | 'Ноутбуки'
    | "Персональні комп'ютери"
    | 'Планшети'
    | 'Смартфони'
    | 'HDD'
    | 'SSD'
    | 'Процесори'
    | 'Відеокарти'
    | "Оперативна пам'ять"
    | 'Монітори'
    | 'Дрони';